rename CurrentSalt to CurrentSaltCommand
