+++ src/commands/CurrentSaltCommand.ts
